requirements.txt
logzero==1.7.0
python-dotenv==1.0.0
web3==5.31.3
